requirements.txt
# Core Dependencies
fastapi==0.121.3
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-core==2.41.5

# AI/ML
openai==2.8.1
anthropic==0.75.0

# HTTP & Networking
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2025.11.12
idna==3.11
sniffio==1.3.1
anyio==4.11.0

# Utilities
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
click==8.3.1
tqdm==4.67.1

# Type Support
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0
docstring-parser==0.17.0

# Server
starlette==0.50.0

# Additional
distro==1.9.0
six==1.17.0
jiter==0.12.0
